ajouts des sous titres

+++ backend/BaseModel/Video.py
@@ -2,7 +2,7 @@ from pydantic import BaseModel
 from typing import List
 
 class VideoMinimumDuration(BaseModel):
-    divide_each_minutes: int
+    divide_each_minutes: float
 
 class VideoStartBefore(BaseModel):
     divide_each_minutes: int
